requirements.txt
python-dotenv
langchain
pydantic
langchain-google-genai
langchain-core
streamlit